reorganize folders to match pub.dev package layout convention
